requirements.txt
boto3
apache-airflow
pyspark
minio
